refactor #1 이미지 컴포넌트 생성
